fix(PyO3): change from_bytes function signature
